Switch to nix developPackage for haskell dependencies
With `developPackage`, nix reads the cabal file and populate the
haskell package database with the required packages already existing
in nixpkgs.

Cabal read the "global" package database before trying to download /
install anything. If satisfying packages exists in the global
database, it won't fetch / build them.

The direct result of this change is that all haskell dependencies of
the project are fetched and cached by nix, meaning a faster startup of
the build.

I've also made the following related changes:

- `shell.nix` renamed to `default.nix`. This does not change the
  behavior of `nix-shell`, but `nix-build` can work without command
  line argument. Hence `.travis.yml` is now simpler because the
  package can be built with `nix-build`.
- The bound for tasty was changed to `< 1.3` (was `< 1.2`) to allow
  the build with `tasty-1.2` which is the version in the selected
  nixpkgs clone.